+++ docs/Doxygen/Chinese.h
@@ -39,11 +39,11 @@
 + 类关系：
   \image html docs/Image/PluginClientAPI.svg
 + 使用：
-  - 实例化类： \ref CManagePlugin m_ManageConnecter;
-    + 处理 CManagePlugin::Handle::onProcess 得到注册的连接插件
+  - 实例化类： \ref CClient m_Client;
+    + 处理 CClient::Handle::onProcess 得到注册的连接插件
   - 使用下面一种方式得到连接者对象：
-    + \ref CManagePlugin::CreateConnecter
-    + \ref CManagePlugin::LoadConnecter
+    + \ref CClient::CreateConnecter
+    + \ref CClient::LoadConnecter
   - 关联信号：
     + 连接成功信号： \ref CConnecter::sigConnected
     + 断连信号： \ref CConnecter::sigDisconnected