+++ ChangeLog.md
@@ -10,6 +10,7 @@
   - Rename CPluginViewerThread to CPluginClientThread
   - Rename CParameterViewer to CParameterClient
   - Rename CConnecter::SetParameterViewer to CConnecter::SetParameterClient
+  - Rename CManagePlugin to CClient
 
 ### v0.0.21
 + Add core dump in windows